commande menu réglé + systéme de commentaire + facture + différentes vues

+++ src/Controller/CommentaireController.php
@@ -5,10 +5,13 @@ namespace App\Controller;
 use App\Entity\Commentaire;
 use App\Form\CommentaireType;
 use App\Repository\CommentaireRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 /**
  * @Route("/commentaire")
@@ -21,31 +24,34 @@ class CommentaireController extends AbstractController
     public function index(CommentaireRepository $commentaireRepository): Response
     {
         return $this->render('commentaire/index.html.twig', [
-            'commentaires' => $commentaireRepository->findAll(),
+            'commentaires' => $commentaireRepository->findBy([], ['commentAt' => 'DESC']),
         ]);
     }
 
     /**
-     * @Route("/new", name="commentaire_new", methods={"GET","POST"})
+     * @Route("/activer/{id}", name="commentaire_activer", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function activer(Request $request, Commentaire $commentaire): Response
     {
-        $commentaire = new Commentaire();
-        $form = $this->createForm(CommentaireType::class, $commentaire);
-        $form->handleRequest($request);
+        $commentaire->setVisible(1);
+        $entityManager = $this->getDoctrine()->getManager();
+        $entityManager->persist($commentaire);
+        $entityManager->flush();
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($commentaire);
-            $entityManager->flush();
+        return $this->redirectToRoute('commentaire_index');
+    }
 
-            return $this->redirectToRoute('commentaire_index');
-        }
+    /**
+     * @Route("/desactiver/{id}", name="commentaire_desactiver", methods={"GET","POST"})
+     */
+    public function desactiver(Request $request, Commentaire $commentaire): Response
+    {
+        $commentaire->setVisible(0);
+        $entityManager = $this->getDoctrine()->getManager();
+        $entityManager->persist($commentaire);
+        $entityManager->flush();
 
-        return $this->render('commentaire/new.html.twig', [
-            'commentaire' => $commentaire,
-            'form' => $form->createView(),
-        ]);
+        return $this->redirectToRoute('commentaire_index');
     }
 
     /**
@@ -79,14 +85,18 @@ class CommentaireController extends AbstractController
     }
 
     /**
+     * @isGranted("ROLE_USER")
      * @Route("/{id}", name="commentaire_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Commentaire $commentaire): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$commentaire->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($commentaire);
-            $entityManager->flush();
+        $user = $this->getUser();
+        if ($user = $commentaire->getUser()) {
+            if ($this->isCsrfTokenValid('delete' . $commentaire->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($commentaire);
+                $entityManager->flush();
+            }
         }
 
         return $this->redirectToRoute('commentaire_index');
